+++ xtask/src/device_path/group.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-use super::node::{is_node_attr, Node};
+use super::node::{Node, is_node_attr};
 use heck::ToUpperCamelCase;
 use proc_macro2::{Span, TokenStream};
 use quote::quote;